Added background loading. Most stuff broken.
I'm re-doing how I display games. I'm going to have a list for Local and
a list for SD games, and switch between them as desired.

However, now it loads games in a background worker so the UI is much
more responsive.
I'm also starting to change the UI to match this, but most of it will
need to be re-done.