No longer useing PackageID
...because some custom XCIs use the same PackageID which causes
problems. Simply using file names now. However, this causes longer load
times.
Upgraded XCI_Explorer to StudentBlake's with some changes to make it
work with SGM.
Compress gameIcons and gameCerts in the cache file on the fly
I don't have the will to work on this much more, and libHac isn't a
simple implementation. If I continue working on this, it will be a
complete rewrite down the road to clean it up.